Remove V3 protocol
