+++ drop-transfer/src/ws/client/mod.rs
@@ -1,6 +1,5 @@
 mod handler;
 mod v2;
-mod v3;
 mod v4;
 
 use std::{
@@ -81,7 +80,6 @@ pub(crate) async fn run(state: Arc<State>, xfer: crate::Transfer, logger: Logger
                 .await
         }
         protocol::Version::V2 => ctx.run(v2::HandlerInit::<true>::new(&state, &logger)).await,
-        protocol::Version::V3 => ctx.run(v3::HandlerInit::new(state, &logger)).await,
         protocol::Version::V4 => ctx.run(v4::HandlerInit::new(state, &logger)).await,
     }
 }
@@ -100,7 +98,6 @@ async fn establish_ws_conn(
 
     let mut versions_to_try = [
         protocol::Version::V4,
-        protocol::Version::V3,
         protocol::Version::V2,
         protocol::Version::V1,
     ]
